+++ src/MEGASync/stalled_issues/gui/stalled_issues_cases/NameConflict.cpp
@@ -80,7 +80,7 @@ void NameConflict::updateUi(std::shared_ptr<const NameConflictedStalledIssue> is
     //Reset widgets
     bool allSolved(true);
 
-    for(int index = conflictedNamesInfo.size()-1; index >= 0; index--)
+    for (int index = static_cast<int>(conflictedNamesInfo.size() - 1); index >= 0; index--)
     {
         std::shared_ptr<NameConflictedStalledIssue::ConflictedNameInfo> info(conflictedNamesInfo.at(index));
         QString conflictedName(getConflictedName(info));
